Backend schema restructure: move logic to service directory . Add new planner logic to service

+++ backend/app/schemas/course.py
@@ -6,10 +6,10 @@ class CourseResponse(BaseModel):
     id: int
     course_name: str
     exam_date: date
-    daily_study_hours: int
+    daily_study_hours: float
     created_at: datetime
 
 class CoursePutRequest(BaseModel):
     course_name: str
     exam_date: date
-    daily_study_hours: int
+    daily_study_hours: float
